New Session - Session Duration Input
Have successfully implemented two EditText inputs specifying the session
duration in minutes and seconds. It now runs a new session using the
duration converted into milliseconds.

Need to next ensure that the inputs are validated appropriately - i.e.,
what happens if the user enters '01' or '1' (then converts it to '01'),
etc?