New Session - Date
Amended setting the session date from purely assigning it to one field,
then storing it to the appropriate session field; to now storing the
date as seperate properties - dayNo, day, month, year.

The Session entity class has been thus modified for this change, in
addition to the SessionHandler class, writing each date property to it.

The session time (the clock time) has been reformatted to include the
time of day - AM/PM.