(Previous) Sessions
SessionsHandler (database) now returns ALL sessions stored, and renders
to the ListView.

Note to self: several setters have been commented out in the
SessionHandler getSessions() method.

There appears to be a bug somewhere when retrieving data.

It might be attributed to the database returning an empty row that is
empty, and the setters doesn't like nulls - it could be an integer
setter returns an empty result or null.