Week (Class)
I have redesigned the Weeks class to incorporate existing private
methods (originally in Sessions.java) that found the beginning and end
of a week.

The class also returns the date as a whole.

I have commented out code (using TODO) that will need to be deleted if
the functionality in the Week's class returns next week's dates tomorrow
(06/09/13).